Safari-609.1.11.1/LayoutTests/svg/animations/animations-paused-in-background-page-iframe-expected.txt

Tests that animated SVG images are properly paused when loaded in a new iframe inside a hidden page

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Setting page visibility to hidden
PASS document.hidden is true
PASS internals.isImageAnimating(testImage) became false
Setting page visibility to visible
PASS internals.isImageAnimating(testImage) became true
PASS successfullyParsed is true

TEST COMPLETE
